Update LAB_AK_02_Lab1_Ex1_IdentitySynchronization.md
Removed static password for Mod Admin, not sure how that got in there.  The only passwords for any of these steps should be Pa55w.rd for on-prem and sync'd users, and Holly.
Updated the removal of users licenses.  My Authorized Lab Hosting Provider does not provision these accounts. Some ALHP do so I modified the step to include "Only complete if the users exist".